Fix incorrect visual location of sneks and food upon scaling window

+++ src/snek/Food.cpp
@@ -1,6 +1,6 @@
 #include "Food.h"
 
-void Food::draw(FoodState *food, int scale) {
-    SDL_Rect rect = { food->pos.x * scale, food->pos.y * scale, scale, scale };
+void Food::draw(FoodState *food, float scale) {
+    SDL_Rect rect = { (int) (food->pos.x * scale), (int) (food->pos.y * scale), (int) scale, (int) scale };
     SDL_RenderFillRect(global::renderer, &rect);
 }
